+++ .storybook/preview.js
@@ -4,9 +4,7 @@ import { ThemeProvider } from 'theme-ui'
 import { theme } from '../theme'
 import { I18nextProvider } from 'react-i18next'
 import i18n from './i18n'
-import { ModalProvider } from 'helpers/modalHook'
 
-addDecorator((storyFn) => <ModalProvider>{storyFn()}</ModalProvider>)
 addDecorator((storyFn) => <I18nextProvider i18n={i18n}>{storyFn()}</I18nextProvider>)
 addDecorator((storyFn) => <ThemeProvider theme={theme}>{storyFn()}</ThemeProvider>)
 addDecorator((story, context) => <Suspense fallback="Loading...">{story(context)}</Suspense>)
