+++ components/vault/VaultDetails.tsx
@@ -1,15 +1,9 @@
 import { Icon } from '@makerdao/dai-ui-icons'
 import BigNumber from 'bignumber.js'
-import { getToken } from 'blockchain/tokensMetadata'
-import { AppLink } from 'components/Links'
 import { Modal, ModalCloseIcon } from 'components/Modal'
-import { formatAmount, formatPercent } from 'helpers/formatters/format'
-import { ModalProps, useModal } from 'helpers/modalHook'
 import { CommonVaultState, WithChildren } from 'helpers/types'
 import { zero } from 'helpers/zero'
-import { Trans } from 'next-i18next'
 import React, { ReactNode } from 'react'
-import { useTranslation } from 'react-i18next'
 import { Box, Card, Flex, Grid, Heading, Text } from 'theme-ui'
 
 type CollRatioColor = 'primary' | 'onError' | 'onWarning' | 'onSuccess'
@@ -71,7 +65,10 @@ export function getAfterPillColors(collRatioColor: CollRatioColor) {
   }
 }
 
-function VaultDetailsAfterPill({ children, afterPillColors }: WithChildren & AfterPillProps) {
+export function VaultDetailsAfterPill({
+  children,
+  afterPillColors,
+}: WithChildren & AfterPillProps) {
   return (
     <Card
       sx={{
@@ -174,566 +171,6 @@ export function VaultDetailsCardModal({
         {children}
       </Grid>
     </Modal>
-  )
-}
-
-export function VaultDetailsCardCurrentPriceModal({
-  close,
-  currentPrice,
-  nextPriceWithChange,
-}: ModalProps<{ currentPrice: ReactNode; nextPriceWithChange: ReactNode }>) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{`${t('manage-multiply-vault.card.current-price')}`}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-multiply-vault.card.current-price-description')}
-        </Text>
-        <Card variant="vaultDetailsCardModal">
-          <Heading variant="header3">{currentPrice}</Heading>
-        </Card>
-      </Grid>
-      <Grid gap={2}>
-        <Heading variant="header3">{`${t('manage-multiply-vault.card.next-price')}`}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {`${t('manage-multiply-vault.card.next-price-description')}`}
-        </Text>
-        <Card variant="vaultDetailsCardModal">
-          <Heading variant="header3">{nextPriceWithChange}</Heading>
-        </Card>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          <Trans
-            i18nKey="manage-multiply-vault.card.more-info-oracles"
-            components={[
-              <AppLink
-                href="https://kb.oasis.app/help/the-oracle-security-module"
-                withAccountPrefix={false}
-                target="_blank"
-                sx={{
-                  display: 'inline-block',
-                  color: 'primary',
-                  textDecoration: 'underline',
-                }}
-              />,
-            ]}
-          />
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-interface CollaterlizationRatioProps {
-  currentCollateralRatio: BigNumber
-  collateralRatioOnNextPrice: BigNumber
-}
-
-export function VaultDetailsCardCollaterlizationRatioModal({
-  currentCollateralRatio,
-  collateralRatioOnNextPrice,
-  close,
-}: ModalProps<CollaterlizationRatioProps>) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{`${t('system.collateralization-ratio')}`}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-vault.card.collateralization-ratio-calculated')}
-        </Text>
-        <Heading variant="header3">
-          {t('manage-vault.card.collateralization-ratio-header2')}
-        </Heading>
-        <Card variant="vaultDetailsCardModal">
-          {formatPercent(currentCollateralRatio.times(100).absoluteValue(), {
-            precision: 2,
-            roundMode: BigNumber.ROUND_DOWN,
-          })}
-        </Card>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-vault.card.collateralization-ratio-description')}
-        </Text>
-        <Heading variant="header3">
-          {t('manage-vault.card.collateralization-ratio-next-price')}
-        </Heading>
-        <Card variant="vaultDetailsCardModal">
-          {formatPercent(collateralRatioOnNextPrice.times(100).absoluteValue(), {
-            precision: 2,
-            roundMode: BigNumber.ROUND_DOWN,
-          })}
-        </Card>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-interface CollateralLockedProps {
-  token: string
-  collateralAmountLocked: BigNumber | undefined
-  collateralLockedUSD: BigNumber | undefined
-}
-
-export function VaultDetailsCardCollateralLockedModal({
-  collateralAmountLocked,
-  collateralLockedUSD,
-  token,
-  close,
-}: ModalProps<CollateralLockedProps>) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{`${t('system.collateral-locked')}`}</Heading>
-        <Heading variant="header3">{`${t('manage-vault.card.collateral-locked-amount')}`}</Heading>
-        <Card variant="vaultDetailsCardModal">
-          {formatAmount(collateralAmountLocked || zero, getToken(token).symbol)}
-        </Card>
-
-        <Heading variant="header3">{t('manage-vault.card.collateral-locked-USD')}</Heading>
-        <Card variant="vaultDetailsCardModal">
-          {collateralLockedUSD && `$${formatAmount(collateralLockedUSD, 'USD')}`}
-        </Card>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-vault.card.collateral-locked-oracles')}
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsBuyingPowerModal({ close }: ModalProps) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{t('manage-multiply-vault.card.buying-power')}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-multiply-vault.card.buying-power-description')}
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsNetValueModal({ close }: ModalProps) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{t('manage-multiply-vault.card.net-value')}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-multiply-vault.card.net-value-description')}
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsStopLossCollRatioModal({ close }: ModalProps) {
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">Stop Loss Coll Ratio</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          StopLossCollRatio dummy modal
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsDynamicStopPriceModal({ close }: ModalProps) {
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">Dynamic Stop Price Modal</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          VaultDetailsDynamicStopPriceModal dummy modal
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsCardMaxTokenOnStopLossTriggerModal({ close }: ModalProps) {
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">Max Token On Stop Loss Trigger</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          VaultDetailsCardMaxTokenOnStopLossTriggerModal dummy modal
-        </Text>
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-interface LiquidationProps {
-  liquidationPrice: BigNumber
-  liquidationPriceCurrentPriceDifference: BigNumber | undefined
-}
-
-export function VaultDetailsLiquidationModal({
-  liquidationPrice,
-  liquidationPriceCurrentPriceDifference,
-  close,
-}: ModalProps<LiquidationProps>) {
-  const { t } = useTranslation()
-  return (
-    <VaultDetailsCardModal close={close}>
-      <Grid gap={2}>
-        <Heading variant="header3">{`${t('system.liquidation-price')}`}</Heading>
-        <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-          {t('manage-multiply-vault.card.liquidation-price-description')}
-        </Text>
-        <Heading variant="header3">
-          {t('manage-multiply-vault.card.liquidation-price-current')}
-        </Heading>
-        <Card variant="vaultDetailsCardModal">{`$${formatAmount(liquidationPrice, 'USD')}`}</Card>
-        {liquidationPriceCurrentPriceDifference && (
-          <Text variant="subheader" sx={{ fontSize: 2, pb: 2 }}>
-            {t(
-              'manage-multiply-vault.card.liquidation-percentage-below',
-              formatPercent(liquidationPriceCurrentPriceDifference.times(100).absoluteValue(), {
-                precision: 2,
-                roundMode: BigNumber.ROUND_DOWN,
-              }),
-            )}
-          </Text>
-        )}
-      </Grid>
-    </VaultDetailsCardModal>
-  )
-}
-
-export function VaultDetailsCardLiquidationPrice({
-  liquidationPrice,
-  liquidationPriceCurrentPriceDifference,
-  afterLiquidationPrice,
-  afterPillColors,
-  showAfterPill,
-}: {
-  liquidationPrice: BigNumber
-  liquidationPriceCurrentPriceDifference?: BigNumber
-  afterLiquidationPrice?: BigNumber
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  return (
-    <VaultDetailsCard
-      title={t('system.liquidation-price')}
-      value={`$${formatAmount(liquidationPrice, 'USD')}`}
-      valueAfter={showAfterPill && `$${formatAmount(afterLiquidationPrice || zero, 'USD')}`}
-      valueBottom={
-        liquidationPriceCurrentPriceDifference && (
-          <>
-            {formatPercent(liquidationPriceCurrentPriceDifference.times(100).absoluteValue(), {
-              precision: 2,
-              roundMode: BigNumber.ROUND_DOWN,
-            })}
-            <Text as="span" sx={{ color: 'text.subtitle' }}>
-              {` ${
-                liquidationPriceCurrentPriceDifference.lt(zero) ? 'above' : 'below'
-              } current price`}
-            </Text>
-          </>
-        )
-      }
-      openModal={() =>
-        openModal(VaultDetailsLiquidationModal, {
-          liquidationPrice: liquidationPrice,
-          liquidationPriceCurrentPriceDifference: liquidationPriceCurrentPriceDifference,
-        })
-      }
-      afterPillColors={afterPillColors}
-    />
-  )
-}
-
-export function VaultDetailsCardCurrentPrice(props: CommonVaultState) {
-  const {
-    priceInfo: {
-      currentCollateralPrice,
-      nextCollateralPrice,
-      isStaticCollateralPrice,
-      collateralPricePercentageChange,
-    },
-  } = props
-  const openModal = useModal()
-  const priceChangeColor = getPriceChangeColor(props)
-
-  const currentPrice = `$${formatAmount(currentCollateralPrice, 'USD')}`
-  const nextPriceWithChange = (
-    <>
-      <Text>${formatAmount(nextCollateralPrice, 'USD')}</Text>
-      <Text sx={{ ml: 2, fontSize: 1 }}>
-        {formatPercent(collateralPricePercentageChange.times(100), { precision: 2 })}
-      </Text>
-    </>
-  )
-
-  return (
-    <VaultDetailsCard
-      title={`Current Price`}
-      value={currentPrice}
-      valueBottom={
-        isStaticCollateralPrice ? null : (
-          <Flex sx={{ whiteSpace: 'pre-wrap' }}>
-            <Text sx={{ color: 'text.subtitle' }}>Next </Text>
-            <Flex
-              variant="paragraph2"
-              sx={{ fontWeight: 'semiBold', alignItems: 'center', color: priceChangeColor }}
-            >
-              {nextPriceWithChange}
-            </Flex>
-          </Flex>
-        )
-      }
-      openModal={() =>
-        openModal(VaultDetailsCardCurrentPriceModal, {
-          currentPrice,
-          nextPriceWithChange: (
-            <Flex sx={{ alignItems: 'center', color: priceChangeColor }}>
-              {nextPriceWithChange}
-            </Flex>
-          ),
-        })
-      }
-    />
-  )
-}
-
-export function VaultDetailsCardCollateralLocked({
-  depositAmountUSD,
-  depositAmount,
-  afterDepositAmountUSD,
-  token,
-  afterPillColors,
-  showAfterPill,
-}: {
-  depositAmountUSD?: BigNumber
-  depositAmount?: BigNumber
-  afterDepositAmountUSD?: BigNumber
-  token: string
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  return (
-    <VaultDetailsCard
-      title={`${t('system.collateral-locked')}`}
-      value={`$${formatAmount(depositAmountUSD || zero, 'USD')}`}
-      valueAfter={showAfterPill && `$${formatAmount(afterDepositAmountUSD || zero, 'USD')}`}
-      valueBottom={
-        <>
-          {formatAmount(depositAmount || zero, getToken(token).symbol)}
-          <Text as="span" sx={{ color: 'text.subtitle' }}>
-            {` ${getToken(token).symbol}`}
-          </Text>
-        </>
-      }
-      openModal={() =>
-        openModal(VaultDetailsCardCollateralLockedModal, {
-          token: token,
-          collateralAmountLocked: depositAmount,
-          collateralLockedUSD: afterDepositAmountUSD,
-        })
-      }
-      afterPillColors={afterPillColors}
-    />
-  )
-}
-
-export function VaultDetailsCardNetValue({
-  netValueUSD,
-  afterNetValueUSD,
-  afterPillColors,
-  showAfterPill,
-}: {
-  netValueUSD: BigNumber
-  afterNetValueUSD: BigNumber
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  return (
-    <VaultDetailsCard
-      title={t('manage-multiply-vault.card.net-value')}
-      value={`$${formatAmount(netValueUSD, 'USD')}`}
-      // valueBottom={`Unrealised P&L 0%`}
-      valueAfter={showAfterPill && `$${formatAmount(afterNetValueUSD, 'USD')}`}
-      openModal={() => openModal(VaultDetailsNetValueModal)}
-      afterPillColors={afterPillColors}
-    />
-  )
-}
-
-export function VaultDetailsCardStopLossCollRatio({
-  slRatio,
-  afterSlRatio,
-  collateralizationRatio,
-  afterPillColors,
-  showAfterPill,
-  isProtected,
-}: {
-  slRatio: BigNumber
-  afterSlRatio?: BigNumber
-  collateralizationRatio: BigNumber
-  isProtected: boolean
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  return (
-    <VaultDetailsCard
-      title={t('manage-multiply-vault.card.stop-loss-coll-ratio')}
-      value={
-        isProtected
-          ? formatPercent(slRatio, {
-              precision: 2,
-            })
-          : '-'
-      }
-      valueBottom={
-        <>
-          {formatPercent(collateralizationRatio, {
-            precision: 2,
-          })}{' '}
-          <Text as="span" sx={{ color: 'text.subtitle' }}>
-            {t('manage-multiply-vault.card.current-coll-ratio')}
-          </Text>
-        </>
-      }
-      valueAfter={
-        showAfterPill &&
-        afterSlRatio &&
-        formatPercent(afterSlRatio, {
-          precision: 2,
-        })
-      }
-      openModal={() => openModal(VaultDetailsStopLossCollRatioModal)}
-      afterPillColors={afterPillColors}
-    />
-  )
-}
-
-export function VaultDetailsCardDynamicStopPrice({
-  slRatio,
-  afterSlRatio,
-  liquidationPrice,
-  afterLiquidationPrice,
-  liquidationRatio,
-  afterPillColors,
-  showAfterPill,
-  isProtected,
-}: {
-  slRatio: BigNumber
-  afterSlRatio?: BigNumber
-  liquidationPrice: BigNumber
-  afterLiquidationPrice?: BigNumber
-  liquidationRatio: BigNumber
-  isProtected: boolean
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  const dynamicStopPrice = liquidationPrice.div(liquidationRatio).times(slRatio)
-  const afterDynamicStopPrice =
-    afterLiquidationPrice && afterSlRatio
-      ? afterLiquidationPrice.div(liquidationRatio).times(afterSlRatio)
-      : zero
-
-  return (
-    <VaultDetailsCard
-      title={t('manage-multiply-vault.card.dynamic-stop-price')}
-      value={isProtected ? `$${formatAmount(dynamicStopPrice, 'USD')}` : '-'}
-      valueBottom={
-        isProtected ? (
-          <>
-            ${formatAmount(dynamicStopPrice.minus(liquidationPrice), 'USD')}{' '}
-            <Text as="span" sx={{ color: 'text.subtitle' }}>
-              {t('manage-multiply-vault.card.above-liquidation-price')}
-            </Text>
-          </>
-        ) : (
-          '-'
-        )
-      }
-      valueAfter={showAfterPill && `$${formatAmount(afterDynamicStopPrice, 'USD')}`}
-      openModal={() => openModal(VaultDetailsDynamicStopPriceModal)}
-      afterPillColors={afterPillColors}
-    />
-  )
-}
-
-export function VaultDetailsCardMaxTokenOnStopLossTrigger({
-  slRatio,
-  afterSlRatio,
-  liquidationPrice,
-  afterPillColors,
-  showAfterPill,
-  isProtected,
-  debt,
-  collateralAmountLocked,
-  afterLockedCollateral,
-  afterDebt,
-  afterLiquidationPrice,
-  liquidationRatio,
-  token,
-}: {
-  slRatio: BigNumber
-  afterSlRatio?: BigNumber
-  liquidationPrice: BigNumber
-  isProtected: boolean
-  debt: BigNumber
-  collateralAmountLocked: BigNumber
-  afterLockedCollateral: BigNumber
-  afterDebt: BigNumber
-  liquidationRatio: BigNumber
-  afterLiquidationPrice: BigNumber
-  token: string
-} & AfterPillProps) {
-  const openModal = useModal()
-  const { t } = useTranslation()
-
-  const ethDuringLiquidation = debt.times(liquidationRatio).div(liquidationPrice)
-
-  const dynamicStopPrice = liquidationPrice.div(liquidationRatio).times(slRatio)
-  const afterDynamicStopPrice = afterLiquidationPrice
-    .div(liquidationRatio)
-    .times(afterSlRatio || zero)
-
-  const maxEth = collateralAmountLocked.times(dynamicStopPrice).minus(debt).div(dynamicStopPrice)
-  const afterMaxEth = afterLockedCollateral
-    .times(afterDynamicStopPrice)
-    .minus(afterDebt)
-    .div(afterDynamicStopPrice)
-
-  return (
-    <VaultDetailsCard
-      title={t('manage-multiply-vault.card.max-token-on-stop-loss-trigger')}
-      value={isProtected ? `${formatAmount(maxEth, token)} ${token}` : '-'}
-      valueBottom={
-        showAfterPill ? (
-          <>
-            ${formatAmount(maxEth.minus(ethDuringLiquidation), token)} {token}{' '}
-            <Text as="span" sx={{ color: 'text.subtitle' }}>
-              {t('manage-multiply-vault.card.saving-comp-to-liquidation')}
-            </Text>
-          </>
-        ) : (
-          '-'
-        )
-      }
-      valueAfter={
-        showAfterPill &&
-        `${t('manage-multiply-vault.card.up-to')} $${formatAmount(afterMaxEth, token)} ${token}`
-      }
-      openModal={() => openModal(VaultDetailsCardMaxTokenOnStopLossTriggerModal)}
-      afterPillColors={afterPillColors}
-    />
   )
 }
